Don't truncate values if they're numbers
The truncate code converts the cell value to string, this ended up
converting numeric values on the sheet to string values.
This commit stops that from happening by first checking if the value
is a number, if it is we don't truncate the value.

A better solution to this is needed. Is truncation even a desirable
feature?